add warning if docker is not found when starting the server in a container

+++ Dev/Run Tests.ps1
@@ -290,11 +290,14 @@ function Cleanup-ServerStudio([bool]$Force=$true) {
     taskkill /im "IEDriverServer.exe" /f  2>&1 | out-null
 
     #Stop Server Container
-    docker exec warewolfserver sc stop "Warewolf Server"
-    $ServerContainerLogText = docker exec warewolfserver cmd /c type "C:\\ProgramData\\Warewolf\\Server Log\\warewolf-server.log"    
-    $ServerContainerLogText | Out-File -LiteralPath "$TestsResultsPath\ServerContainer.log" -Encoding utf8 -Force
-    docker stop warewolfserver
-    docker container rm -f warewolfserver
+    if (Get-Command docker -ErrorAction SilentlyContinue) {
+        docker exec warewolfserver sc stop "Warewolf Server"
+        $ServerContainerLogText = docker exec warewolfserver cmd /c type "C:\\ProgramData\\Warewolf\\Server Log\\warewolf-server.log"    
+        $ServerContainerLogText | Out-File -LiteralPath "$TestsResultsPath\ServerContainer.log" -Encoding utf8 -Force
+        docker stop warewolfserver
+        docker container rm -f warewolfserver
+        docker images | ConvertFrom-String | where {$_.P2 -eq "<none>"} | % { docker rmi $_.P3 }
+    }
 
     #Delete Certain Studio and Server Resources
     $ToClean = "$env:LOCALAPPDATA\Warewolf\DebugData\PersistSettings.dat",
@@ -646,18 +649,22 @@ function Start-Server([string]$ServerPath,[string]$ResourcesType) {
 }
 
 function Start-Container {
-    if ($ServerPath -eq "" -or !(Test-Path $ServerPath)) {
-        $ServerPath = Find-Warewolf-Server-Exe
+    if (Get-Command docker -ErrorAction SilentlyContinue) {
+        if ($ServerPath -eq "" -or !(Test-Path $ServerPath)) {
+            $ServerPath = Find-Warewolf-Server-Exe
+        }
+        $ServerDirectory = (Get-Item $ServerPath).Directory.FullName
+        Write-Host Starting container from $ServerDirectory
+        $ErrorActionPreference = 'silentlycontinue'
+        docker network create "My Container Network"
+        docker container rm -f warewolfserver
+        $ErrorActionPreference = 'Continue'
+        docker build -t warewolfserver "$ServerDirectory"
+        docker run --name warewolfserver --hostname localwarewolfservercontainer --network "My Container Network" -d warewolfserver ping -t 4.2.2.3
+        Write-Host Server container has started.
+    } else {
+        Write-Warning -Message "Cannot find Docker, container server not started."
     }
-    $ServerDirectory = (Get-Item $ServerPath).Directory.FullName
-    Write-Host Starting container from $ServerDirectory
-    $ErrorActionPreference = 'silentlycontinue'
-    docker network create "My Container Network"
-    docker container rm -f warewolfserver
-    $ErrorActionPreference = 'Continue'
-    docker build -t warewolfserver "$ServerDirectory"
-    docker run --name warewolfserver --hostname localwarewolfservercontainer --network "My Container Network" -d warewolfserver ping -t 4.2.2.3
-    Write-Host Server container has started.
 }
 
 function Start-my.warewolf.io {
